Added code for project
Created a new message for completing quiz showing the correct number of
first guess. INstead of creating the bonus questions as i did not want
to create an array of capitals, i created a new pref to set the number
of flags for the quiz, however there was an issue with the file path for
this project causing the app to crash.